# Bug/66405 poor performance on workpackage visible scope