# Ticket

https://community.openproject.org/wp/66405

# What are you trying to accomplish?

Attempting to improve the performance on all the places where `WorkPackages.allowed_to` is called. There are quite a few among them:
* The check for whether a timer is running to display that information in the avatar at the top right corner. This is checked every time the menu is rendered so on every page refresh.
* Fetching the visible work packages to display in the work package table/board/team planner.
* Fetching candidates for the work package autocompleter when setting relations
* Fetching candidates for the work package autocompleter when typing in a comment field in the wp activity
* Fetching notifications

Since not all installations have the same data structure and the query shows good performance metrics on most, a setting is introduced that allows switching to an SQL structure more fit for larger installations (projects, users, members & work packages). In the past, when trying to optimize for larger instances, smaller ones had their performance decreased. With the setting, the best approach can be decided on after deployment.

On the large instance used for testing, the gains are noticeable:
* The times for a request on the root page are roughly halved from 500ms to 270ms.
* The times on a request for 200 WPs via the API drops from 5s to 3s
* The times for autocompletion on typing in a comment field drops from 600ms to 120ms.

However, there are also used cases where the performance is not improved 
* Fetching candidates for the work package autocompleter when setting relations
* Fetching notifications
But the performance on those is also not decreased.

The fix only addresses "normal users", meaning that neither the behaviour for admins nor the one for anonymous is changed. 

# Merge checklist

- [x] Added/updated tests
- [ ] Added/updated documentation in Lookbook (patterns, previews, etc)
- [ ] Tested major browsers (Chrome, Firefox, Edge, ...)